requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
psycopg2==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.2.1
sqlparse==0.5.5
uritemplate==4.2.0
